refactor(nimbus): move all results processing methods and tests into separate files
