+++ experimenter/experimenter/experiments/tests/test_models.py
@@ -56,12 +56,8 @@ from experimenter.experiments.tests.factories import (
 from experimenter.experiments.tests.jexl_utils import validate_jexl_expr
 from experimenter.features import Features
 from experimenter.features.tests import mock_valid_features
-from experimenter.metrics import MetricAreas
-from experimenter.metrics.tests import mock_valid_metrics
 from experimenter.nimbus_ui.constants import NimbusUIConstants
 from experimenter.openidc.tests.factories import UserFactory
-from experimenter.outcomes import Outcomes
-from experimenter.outcomes.tests import mock_valid_outcomes
 from experimenter.projects.tests.factories import ProjectFactory
 
 
@@ -1447,176 +1443,6 @@ class TestNimbusExperiment(TestCase):
             self.assertNotIn(ai_targeting_expr, experiment.targeting)
         validate_jexl_expr(experiment.targeting, experiment.application)
 
-    @mock_valid_outcomes
-    def test_get_weekly_metric_data(self):
-        application = NimbusExperiment.Application.DESKTOP
-        Outcomes.clear_cache()
-        desktop_outcome_1 = Outcomes.get_by_slug_and_application(
-            "desktop_outcome_1", application
-        )
-        experiment = NimbusExperimentFactory.create(
-            application=NimbusExperiment.Application.DESKTOP,
-            primary_outcomes=[desktop_outcome_1.slug],
-            secondary_outcomes=[],
-        )
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-        branch_b = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch B", slug="branch-b"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "weekly": {
-                    "enrollments": {
-                        "all": {
-                            branch_a.slug: {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "urlbar_amazon_search_count": {
-                                            "absolute": {
-                                                "all": [
-                                                    {
-                                                        "lower": 140,
-                                                        "upper": 160,
-                                                        "point": 150,
-                                                    },
-                                                    {
-                                                        "lower": 130,
-                                                        "upper": 150,
-                                                        "point": 140,
-                                                    },
-                                                    {
-                                                        "lower": 120,
-                                                        "upper": 140,
-                                                        "point": 130,
-                                                    },
-                                                ]
-                                            },
-                                            "relative_uplift": {
-                                                branch_a.slug: {"all": []},
-                                            },
-                                        }
-                                    }
-                                },
-                            },
-                            branch_b.slug: {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "urlbar_amazon_search_count": {
-                                            "absolute": {
-                                                "all": [
-                                                    {
-                                                        "lower": 140,
-                                                        "upper": 160,
-                                                        "point": 150,
-                                                    },
-                                                    {
-                                                        "lower": 130,
-                                                        "upper": 150,
-                                                        "point": 140,
-                                                    },
-                                                    {
-                                                        "lower": 120,
-                                                        "upper": 140,
-                                                        "point": 130,
-                                                    },
-                                                ]
-                                            },
-                                            "relative_uplift": {
-                                                branch_a.slug: {
-                                                    "all": [
-                                                        {
-                                                            "lower": 10,
-                                                            "upper": 20,
-                                                            "point": 15,
-                                                        },
-                                                        {
-                                                            "lower": 5,
-                                                            "upper": 15,
-                                                            "point": 10,
-                                                        },
-                                                        {
-                                                            "lower": 0,
-                                                            "upper": 10,
-                                                            "point": 5,
-                                                        },
-                                                    ]
-                                                },
-                                                branch_b.slug: {"all": []},
-                                            },
-                                        }
-                                    }
-                                }
-                            },
-                        }
-                    }
-                }
-            }
-        }
-        experiment.save()
-
-        expected_weekly_data = {
-            "urlbar_amazon_search_count": {
-                "data": {
-                    branch_a.slug: [
-                        (
-                            {"lower": 140, "upper": 160, "significance": "neutral"},
-                            None,
-                        ),
-                        (
-                            {"lower": 130, "upper": 150, "significance": "neutral"},
-                            None,
-                        ),
-                        (
-                            {"lower": 120, "upper": 140, "significance": "neutral"},
-                            None,
-                        ),
-                    ],
-                    branch_b.slug: [
-                        (
-                            {"lower": 140, "upper": 160, "significance": "neutral"},
-                            {
-                                "avg_rel_change": 15,
-                                "lower": 10,
-                                "upper": 20,
-                                "significance": "neutral",
-                            },
-                        ),
-                        (
-                            {"lower": 130, "upper": 150, "significance": "neutral"},
-                            {
-                                "avg_rel_change": 10,
-                                "lower": 5,
-                                "upper": 15,
-                                "significance": "neutral",
-                            },
-                        ),
-                        (
-                            {"lower": 120, "upper": 140, "significance": "neutral"},
-                            {
-                                "avg_rel_change": 5,
-                                "lower": 0,
-                                "upper": 10,
-                                "significance": "neutral",
-                            },
-                        ),
-                    ],
-                },
-                "has_weekly_data": True,
-            },
-            "total_amazon_search_count": {
-                "data": {},
-                "has_weekly_data": False,
-            },
-        }
-
-        self.assertEqual(
-            experiment.get_weekly_metric_data("enrollments", "all", branch_a.slug),
-            expected_weekly_data,
-        )
-
     @parameterized.expand(
         [
             [
@@ -2585,578 +2411,6 @@ class TestNimbusExperiment(TestCase):
                         link["active"],
                         f"{link['title']} should not be active for path {path}",
                     )
-
-    def test_get_branch_data_returns_correct_data(self):
-        experiment = NimbusExperimentFactory.create()
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-        branch_b = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch B", slug="branch-b"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "overall": {
-                    "enrollments": {
-                        "all": {
-                            "branch-a": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "identity": {
-                                            "absolute": {"first": {"point": 150}},
-                                            "percent": 12,
-                                        }
-                                    }
-                                }
-                            },
-                            "branch-b": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "identity": {
-                                            "absolute": {"first": {"point": 75}},
-                                            "percent": 88,
-                                        }
-                                    }
-                                }
-                            },
-                        }
-                    }
-                }
-            }
-        }
-        experiment.save()
-
-        result = experiment.get_branch_data("enrollments", "all")
-
-        self.assertEqual(len(result), 4)
-
-        index_map = {item.get("slug"): i for i, item in enumerate(result)}
-        first = result[index_map.get("branch-a")]
-        second = result[index_map.get("branch-b")]
-
-        # Validate first branch
-        self.assertEqual(first["slug"], "branch-a")
-        self.assertEqual(first["name"], "Branch A")
-        self.assertEqual(first["percentage"], 12)
-        self.assertEqual(first["num_participants"], 150)
-        self.assertEqual(first["description"], branch_a.description)
-
-        # Validate second branch
-        self.assertEqual(second["slug"], "branch-b")
-        self.assertEqual(second["name"], "Branch B")
-        self.assertEqual(second["percentage"], 88)
-        self.assertEqual(second["num_participants"], 75)
-        self.assertEqual(second["description"], branch_b.description)
-
-    @mock_valid_outcomes
-    def test_get_metric_data_returns_correct_data(self):
-        application = NimbusExperiment.Application.DESKTOP
-        Outcomes.clear_cache()
-        desktop_outcome_1 = Outcomes.get_by_slug_and_application(
-            "desktop_outcome_1", application
-        )
-        experiment = NimbusExperimentFactory.create(
-            application=NimbusExperiment.Application.DESKTOP,
-            primary_outcomes=[desktop_outcome_1.slug],
-            secondary_outcomes=[],
-        )
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-        branch_b = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch B", slug="branch-b"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "overall": {
-                    "enrollments": {
-                        "all": {
-                            "branch-a": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "urlbar_amazon_search_count": {
-                                            "absolute": {
-                                                "all": [
-                                                    {
-                                                        "lower": 1.49,
-                                                        "upper": 1.74,
-                                                        "point": 1.62,
-                                                    }
-                                                ]
-                                            },
-                                            "relative_uplift": {
-                                                "branch-a": {"all": []},
-                                                "branch-b": {
-                                                    "all": [
-                                                        {
-                                                            "lower": -0.12,
-                                                            "upper": 0.15,
-                                                            "point": 0.02,
-                                                        }
-                                                    ]
-                                                },
-                                            },
-                                            "significance": {
-                                                "branch-a": {"overall": {}},
-                                                "branch-b": {"overall": {"1": "neutral"}},
-                                            },
-                                            "difference": {
-                                                "branch-a": {"all": []},
-                                                "branch-b": {
-                                                    "all": [
-                                                        {
-                                                            "lower": 0.01,
-                                                            "upper": 0.03,
-                                                            "point": 0.02,
-                                                        }
-                                                    ]
-                                                },
-                                            },
-                                        }
-                                    }
-                                }
-                            },
-                            "branch-b": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "urlbar_amazon_search_count": {
-                                            "absolute": {
-                                                "all": [
-                                                    {
-                                                        "lower": 1.24,
-                                                        "upper": 1.63,
-                                                        "point": 1.43,
-                                                    }
-                                                ]
-                                            },
-                                            "relative_uplift": {
-                                                "branch-a": {"all": [{}]},
-                                                "branch-b": {"all": []},
-                                            },
-                                            "significance": {
-                                                "branch-a": {
-                                                    "overall": {"1": "positive"}
-                                                },
-                                                "branch-b": {"overall": {}},
-                                            },
-                                            "difference": {
-                                                "branch-a": {
-                                                    "all": [
-                                                        {
-                                                            "lower": 0.01,
-                                                            "upper": 0.03,
-                                                            "point": 0.02,
-                                                        }
-                                                    ]
-                                                },
-                                                "branch-b": {"all": []},
-                                            },
-                                        }
-                                    }
-                                }
-                            },
-                        }
-                    }
-                }
-            }
-        }
-        experiment.save()
-
-        results_a = experiment.get_metric_data("enrollments", "all", "branch-a")
-        results_b = experiment.get_metric_data("enrollments", "all", "branch-b")
-
-        data_a = (
-            results_b.get(desktop_outcome_1.friendly_name, {})
-            .get("data", {})
-            .get("overall", {})
-        )
-        data_b = (
-            results_a.get(desktop_outcome_1.friendly_name, {})
-            .get("data", {})
-            .get("overall", {})
-        )
-
-        self.assertEqual(
-            data_a.get("urlbar_amazon_search_count").get(branch_a.slug).get("absolute"),
-            [{"lower": 1.49, "upper": 1.74, "significance": "neutral"}],
-        )
-        self.assertEqual(
-            data_a.get("urlbar_amazon_search_count").get(branch_a.slug).get("relative"),
-            [
-                {
-                    "lower": -0.12,
-                    "upper": 0.15,
-                    "significance": "neutral",
-                    "avg_rel_change": 0.02,
-                }
-            ],
-        )
-
-        self.assertEqual(
-            data_b.get("urlbar_amazon_search_count").get(branch_b.slug).get("absolute"),
-            [{"lower": 1.24, "upper": 1.63, "significance": "positive"}],
-        )
-
-    @mock_valid_metrics
-    def test_metric_areas_created_correctly(self):
-        experiment = NimbusExperimentFactory.create(
-            application=NimbusExperiment.Application.DESKTOP,
-        )
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "other_metrics": {
-                    "other_metrics": {
-                        "mock_engagement_metric": "Metric Name",
-                    }
-                },
-            }
-        }
-        experiment.save()
-
-        MetricAreas.clear_cache()
-        metric_areas = experiment.get_metric_areas(
-            "enrollments", "all", branch_a.slug
-        ).keys()
-
-        self.assertIn("KPI Metrics", metric_areas)
-        self.assertIn("Engagement", metric_areas)
-
-    @parameterized.expand(
-        [
-            (
-                "client_level_daily_active_users_v2",
-                [
-                    {
-                        "group": "other_metrics",
-                        "friendly_name": "Retention",
-                        "slug": "retained",
-                        "description": "Percentage of users who returned to Firefox two weeks later.",  # noqa E501
-                        "display_type": "percentage",
-                    },
-                    {
-                        "group": "search_metrics",
-                        "friendly_name": "Search Count",
-                        "slug": "search_count",
-                        "description": "Daily mean number of searches per user.",
-                    },
-                    {
-                        "group": "other_metrics",
-                        "friendly_name": "Daily Active Users",
-                        "slug": "client_level_daily_active_users_v2",
-                        "description": "Average number of client that sent a main ping per day.",  # noqa E501
-                    },
-                ],
-            ),
-            (
-                "days_of_use",
-                [
-                    {
-                        "group": "other_metrics",
-                        "friendly_name": "Retention",
-                        "slug": "retained",
-                        "description": "Percentage of users who returned to Firefox two weeks later.",  # noqa E501
-                        "display_type": "percentage",
-                    },
-                    {
-                        "group": "search_metrics",
-                        "friendly_name": "Search Count",
-                        "slug": "search_count",
-                        "description": "Daily mean number of searches per user.",
-                    },
-                    {
-                        "group": "other_metrics",
-                        "friendly_name": "Days of Use",
-                        "slug": "days_of_use",
-                        "description": "Average number of days each client sent a main ping.",  # noqa E501
-                    },
-                ],
-            ),
-        ]
-    )
-    def test_get_kpi_metrics_returns_correct_metrics(
-        self, kpi_slug, expected_kpi_metrics
-    ):
-        experiment = NimbusExperimentFactory.create()
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "overall": {
-                    "enrollments": {
-                        "all": {
-                            "branch-a": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        kpi_slug: {
-                                            "absolute": {
-                                                "all": [
-                                                    {
-                                                        "lower": 1.19,
-                                                        "upper": 1.74,
-                                                        "point": 1.62,
-                                                    }
-                                                ]
-                                            },
-                                            "relative_uplift": {
-                                                "branch-a": {"all": []},
-                                                "branch-b": {
-                                                    "all": [
-                                                        {
-                                                            "lower": -0.1,
-                                                            "upper": 0.42,
-                                                            "point": 0.2,
-                                                        }
-                                                    ]
-                                                },
-                                            },
-                                            "significance": {
-                                                "branch-a": {"overall": {}},
-                                                "branch-b": {"overall": {"1": "neutral"}},
-                                            },
-                                            "difference": {
-                                                "branch-a": {"all": []},
-                                                "branch-b": {
-                                                    "all": [
-                                                        {
-                                                            "lower": 0.01,
-                                                            "upper": 0.03,
-                                                            "point": 0.02,
-                                                        }
-                                                    ]
-                                                },
-                                            },
-                                        }
-                                    }
-                                }
-                            },
-                        }
-                    }
-                }
-            }
-        }
-
-        experiment.save()
-        kpi_metrics = experiment.get_kpi_metrics("enrollments", "all", branch_a.slug)
-
-        self.assertListEqual(kpi_metrics, expected_kpi_metrics)
-
-    def test_get_defaults_metrics_with_exclusions(self):
-        experiment = NimbusExperimentFactory.create(
-            results_data={
-                "v3": {
-                    "metadata": {
-                        "metrics": {
-                            "metricA": {
-                                "retained": "Retained",
-                                "search_count": "Search Count",
-                            }
-                        },
-                    },
-                    "other_metrics": {
-                        "other_metrics": {
-                            "retained": "2 Week Retention",
-                            "search_count": "Search Count",
-                        }
-                    },
-                }
-            }
-        )
-
-        remaining_metrics = experiment.get_remaining_metrics_metadata(
-            exclude_slugs=["search_count"]
-        )
-        metric_slugs = [metric.get("slug") for metric in remaining_metrics]
-
-        self.assertIn("retained", metric_slugs)
-        self.assertNotIn("search_count", metric_slugs)
-
-    @parameterized.expand(
-        [
-            (
-                {
-                    "v3": {
-                        "overall": {
-                            "enrollments": {
-                                "all": {
-                                    "branch-a": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {"all": []},
-                                                        "branch-b": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -0.12,
-                                                                    "upper": 0.15,
-                                                                    "point": 0.02,
-                                                                }
-                                                            ]
-                                                        },
-                                                        "branch-c": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -0.1,
-                                                                    "upper": 0.2,
-                                                                    "point": 0.03,
-                                                                }
-                                                            ]
-                                                        },
-                                                    },
-                                                }
-                                            }
-                                        }
-                                    },
-                                    "branch-b": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -2.3,
-                                                                    "upper": 2.1,
-                                                                    "point": 1.13,
-                                                                }
-                                                            ]
-                                                        },
-                                                        "branch-b": {"all": []},
-                                                        "branch-c": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -0.25,
-                                                                    "upper": 0.45,
-                                                                    "point": 0.1,
-                                                                }
-                                                            ]
-                                                        },
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    },
-                                    "branch-c": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -0.3,
-                                                                    "upper": 1.68,
-                                                                    "point": 1.458,
-                                                                }
-                                                            ]
-                                                        },
-                                                        "branch-b": {
-                                                            "all": [
-                                                                {
-                                                                    "lower": -0.25,
-                                                                    "upper": 0.45,
-                                                                    "point": 0.1,
-                                                                }
-                                                            ]
-                                                        },
-                                                        "branch-c": {"all": []},
-                                                    },
-                                                }
-                                            }
-                                        }
-                                    },
-                                }
-                            }
-                        }
-                    }
-                },
-                2.3,
-            ),
-            (
-                {
-                    "v3": {
-                        "overall": {
-                            "enrollments": {
-                                "all": {
-                                    "branch-a": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {"all": []},
-                                                        "branch-b": {"all": [{}]},
-                                                        "branch-c": {"all": [{}]},
-                                                    },
-                                                }
-                                            }
-                                        }
-                                    },
-                                    "branch-b": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {"all": [{}]},
-                                                        "branch-b": {"all": []},
-                                                        "branch-c": {"all": [{}]},
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    },
-                                    "branch-c": {
-                                        "branch_data": {
-                                            "other_metrics": {
-                                                "retained": {
-                                                    "relative_uplift": {
-                                                        "branch-a": {"all": [{}]},
-                                                        "branch-b": {"all": [{}]},
-                                                        "branch-c": {"all": []},
-                                                    },
-                                                }
-                                            }
-                                        }
-                                    },
-                                }
-                            }
-                        }
-                    }
-                },
-                0,
-            ),
-        ]
-    )
-    def test_get_max_metric_value(self, results_data, expected_max_value):
-        experiment = NimbusExperimentFactory.create()
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-        branch_b = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch B", slug="branch-b"
-        )
-        branch_c = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch C", slug="branch-c"
-        )
-
-        experiment.results_data = results_data
-
-        branch_a.save()
-        branch_b.save()
-        branch_c.save()
-        experiment.save()
-
-        extreme_retained = experiment.get_max_metric_value(
-            "enrollments", "all", "branch-a", "other_metrics", "retained"
-        )
-        self.assertEqual(extreme_retained, expected_max_value)
 
     def test_conclusion_recommendation_labels(self):
         recommendations = list(NimbusConstants.ConclusionRecommendation)
@@ -5829,55 +5083,6 @@ class TestNimbusExperiment(TestCase):
         experiment.save()
 
         self.assertEqual(experiment.has_results_errors, expected_results)
-
-    def test_experiment_kpi_metrics_have_errors(self):
-        experiment = NimbusExperimentFactory.create()
-        branch_a = NimbusBranchFactory.create(
-            experiment=experiment, name="Branch A", slug="branch-a"
-        )
-
-        experiment.results_data = {
-            "v3": {
-                "errors": {
-                    "client_level_daily_active_users_v2": [
-                        {"analysis_basis": "enrollments", "segment": "all"}
-                    ],
-                    "experiment": [],
-                },
-                "overall": {
-                    "enrollments": {
-                        "all": {
-                            "branch-a": {
-                                "branch_data": {
-                                    "other_metrics": {
-                                        "client_level_daily_active_users_v2": {
-                                            "absolute": {"all": []},
-                                            "difference": {
-                                                "branch-a": {"all": [{}]},
-                                            },
-                                        }
-                                    }
-                                }
-                            },
-                        }
-                    }
-                },
-            }
-        }
-
-        experiment.save()
-        kpi_metrics = experiment.get_kpi_metrics("enrollments", "all", branch_a.slug)
-
-        self.assertIn(
-            {
-                "group": "other_metrics",
-                "friendly_name": "Daily Active Users",
-                "slug": "client_level_daily_active_users_v2",
-                "description": "Average number of client that sent a main ping per day.",
-                "has_errors": True,
-            },
-            kpi_metrics,
-        )
 
 
 class TestNimbusBranch(TestCase):
